% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafunctions.R
\name{create_email_as_json}
\alias{create_email_as_json}
\title{Convert Email Data to JSON}
\usage{
create_email_as_json(
  data = NULL,
  sort.by.decreasing.datetime = FALSE,
  rename_duplicate_id = TRUE,
  addmainbrackets = TRUE,
  savejson = TRUE,
  file_path = NULL,
  catdata = FALSE,
  writewith = "file_conn",
  info = TRUE,
  debug_info = FALSE
)
}
\arguments{
\item{data}{A list containing email or any kind of list data in R that you want to convert into json-format.}

\item{sort.by.decreasing.datetime}{sort.by.decreasing.datetime default NULL, TRUE = desc i.e. decreasing, FALSE = asc i.e. incresing.}

\item{rename_duplicate_id}{rename_duplicate_id if duplicate id values, these are corrected automatically. Note, it is recommend not to have duplicate id values, otherwise linking between objects may not work.}

\item{addmainbrackets}{Add start and end brackets to the output. Default is TRUE.}

\item{savejson}{Do you want to save json, default is TRUE.}

\item{file_path}{The full path to the file. Remember to include \emph{.json}.}

\item{catdata}{print output in the R Console. Default FALSE.}

\item{writewith}{writewith defines encoding, default is "file_conn" using UTF-8.}

\item{info}{show informative output on Console. Default: TRUE}

\item{debug_info}{show debug output on Console. Default: FALSE}
}
\value{
A JSON string representing the email data without brackets around single elements.
}
\description{
This function converts email data stored in a list to JSON format.
}
\section{Changes in version 0.02.14}{

\itemize{
\item json output comma bug fixed, it was missing between objects in the generated array.
\item sort by datetime and process creates \emph{posixct} key to each object. If not NULL: \emph{sort.by.decreasing.datetime}.
\item duplicate rename process, see argument: \emph{rename_duplicate_id}.
\item input argument \emph{emails} changed to more general \emph{data}.
}
}

\section{Changes in version 0.02.12}{

\itemize{
\item UTF-8 encoding
}
}

\section{Changes in version 0.02.11}{

\itemize{
\item Initial release
}
}

\examples{
emails <- list(
  date = "2024-02-11T09:30:00",
  sender = "john.doe@example.com",
  receiver = "jane.smith@example.com",
  subject = "Re: Follow-up",
  content = "Hi Jane,\n\nJust following up on our meeting yesterday.
   Let me know if you need anything.\n\nBest regards,\nJohn",
  notes = "",
  keywords = c("follow-up", "meeting", "thanks")
)
\dontrun{
 create_email_as_json(data = emails)
}
}
