% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafunctions.R
\name{create_email_as_json}
\alias{create_email_as_json}
\title{Convert Email Data to JSON}
\usage{
create_email_as_json(
  emails = NULL,
  addmainbrackets = TRUE,
  savejson = TRUE,
  file_path = NULL,
  info = FALSE
)
}
\arguments{
\item{emails}{A list containing email data.}

\item{addmainbrackets}{addmainbrackets}

\item{savejson}{savejson}

\item{file_path}{file_path}

\item{info}{info}
}
\value{
A JSON string representing the email data without brackets around single elements.
}
\description{
This function converts email data stored in a list to JSON format.
}
\examples{
emails <- list(
  date = "2024-02-11T09:30:00",
  sender = "john.doe@example.com",
  receiver = "jane.smith@example.com",
  subject = "Re: Follow-up",
  content = "Hi Jane,\n\nJust following up on our meeting yesterday.
   Let me know if you need anything.\n\nBest regards,\nJohn",
  notes = "",
  keywords = c("follow-up", "meeting", "thanks")
)
\dontrun{
 create_email_as_json(emails)
}
}
